Skeletal Watch app commit
If I ever figure out how to do drawing on watch, the app’s UI is sort of ready. A drawing view that will give a popup listing the commands found, and a browsing view that will have a search button. The latter is secondary - a drawing view is the most important part of the app