Major changes in how views are defined by state
Views are defined by state in SwiftUI. In my app, state is defined by the instance of a class LabelScores - labelScores. labelScores has a value associated with it only when CanvasView has a drawing, and it has been recognized by the CoreML model.

All changes in this commit reflect this.