Catalyst version is much simpler now
The Mac Catalyst version is never going to look great. This is fine - Its no longer a 2 column main view. The search field is ugly, and the row spacing is little off, but there's not much I can do to fix it. Its functional and works, and that's all Mac Catalyst apps are set out to do.